Move servo code to modules/servo.*
